requirements.txt
flask
python-docx
num2words
python-telegram-bot
python-dotenv
python-dateutil